fixed value entity

+++ ServiceModule/Thresholds/DataAccess/Entities/Value.cs
@@ -5,10 +5,10 @@ namespace ServiceModule.Thresholds.DataAccess.Entities
     public class Value
     {
         [JsonProperty("more")]
-        public int More { get; set; }
+        public float More { get; set; }
 
         [JsonProperty("less")]
-        public int Less { get; set; }
+        public float Less { get; set; }
 
         [JsonProperty("equal")]
         public string Equal { get; set; }
